auth middleware

+++ src/route/experience.js
@@ -3,13 +3,11 @@ const router = express.Router();
 const User = require("../model/user");
 const Experience = require("../model/experience");
 const { check, validationResult } = require("express-validator");
+const { basic } = require("../utils/auth");
 
 router.post(
-  "/:email",
+  "/",
   [
-    check("email")
-      .isEmail()
-      .withMessage("A valid email is required!"),
     check("role")
       .isLength({ min: 2 })
       .withMessage("Role is required. Min. Lenght is 2 characters."),
@@ -18,13 +16,15 @@ router.post(
       .withMessage("Company is required. Min. Lenght is 2 characters."),
     check("startDate").isISO8601({ strict: true })
   ],
+  basic,
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(422).json({ errors: errors.array() });
     }
-    const email = req.params.email;
     try {
+      // getting username from auth middleware
+      const username = req.auth.user;
       const { role, company, startDate, endDate, description, area } = req.body;
       let newExperience = { role, company, startDate };
       if (endDate) newExperience = { ...newExperience, endDate };
@@ -33,7 +33,7 @@ router.post(
       const response = await Experience.create(newExperience);
       await response.save();
       await User.findOneAndUpdate(
-        { email },
+        { username },
         {
           $push: {
             experiences: response._id
@@ -59,22 +59,35 @@ router.put(
       .withMessage("Company is required. Min. Lenght is 2 characters."),
     check("startDate").isISO8601({ strict: true })
   ],
+  basic,
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(422).json({ errors: errors.array() });
     }
     try {
+      const username = req.auth.user;
       const _id = req.params._id;
-      const { role, company, startDate, endDate, description, area } = req.body;
-      let update = { role, company, startDate };
-      if (endDate) update = { ...update, endDate };
-      if (description) update = { ...update, description };
-      if (area) update = { ...update, area };
-      const response = await Experience.findByIdAndUpdate(_id, update, {
-        new: true
-      });
-      response ? res.json(response) : res.status(404).json({});
+      const user = await User.findOne({ username });
+      const isUserExp = user.experiences.find(exp => exp.toString() === _id);
+      if (isUserExp) {
+        const {
+          role,
+          company,
+          startDate,
+          endDate,
+          description,
+          area
+        } = req.body;
+        let update = { role, company, startDate };
+        if (endDate) update = { ...update, endDate };
+        if (description) update = { ...update, description };
+        if (area) update = { ...update, area };
+        const response = await Experience.findByIdAndUpdate(_id, update, {
+          new: true
+        });
+        res.json(response);
+      } else res.status(404).json({});
     } catch (error) {
       console.log(error);
       res.json(error);
@@ -82,11 +95,16 @@ router.put(
   }
 );
 
-router.delete("/:_id", async (req, res) => {
-  const _id = req.params._id;
+router.delete("/:_id", basic, async (req, res) => {
   try {
-    const response = await Experience.findByIdAndDelete(_id);
-    response ? res.json(response) : res.status(404).json({});
+    const username = req.auth.user;
+    const _id = req.params._id;
+    const user = await User.findOne({ username });
+    const isUserExp = user.experiences.find(exp => exp.toString() === _id);
+    if (isUserExp) {
+      const response = await Experience.findByIdAndDelete(_id);
+      res.json(response);
+    } else res.status(404).json({});
   } catch (error) {
     console.log(error);
     res.json(error);
